requirements.txt
flask
deta
python-dotenv
